+++ source/Public/Remove-SqlElasticJob.ps1
@@ -101,7 +101,7 @@ function Remove-SqlElasticJob
                 return
             }
 
-            Write-PSFMessage -Level Verbose -Message ('{0} Nothing to remove.' -f $message) -Tag 'idempotent'
+            Write-PSFMessage -Level Output -Message ('{0} Nothing to remove.' -f $message) -Tag 'idempotent'
 
             return
         }
@@ -120,11 +120,11 @@ function Remove-SqlElasticJob
 
         $removeParameters = Add-OptionalParameter -Parameter $removeParameters -BoundParameter $PSBoundParameters -Name 'Force'
 
-        Write-PSFMessage -Level Verbose -Message ('Removing Elastic Job ''{0}'' from agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'remove'
+        Write-PSFMessage -Level Output -Message ('Removing Elastic Job ''{0}'' from agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'remove'
 
         $null = Remove-AzSqlElasticJob @removeParameters
 
-        Write-PSFMessage -Level Verbose -Message ('Removed Elastic Job ''{0}'' from agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'remove'
+        Write-PSFMessage -Level Output -Message ('Removed Elastic Job ''{0}'' from agent ''{1}''.' -f $Name, $AgentName) -Tag 'job', 'remove'
 
         if ($PassThru.IsPresent)
         {