+++ source/Public/Test-SqlElasticJobEnvironment.ps1
@@ -76,7 +76,7 @@ function Test-SqlElasticJobEnvironment
             $agent = Get-SqlElasticJobAgent -ResourceGroupName $ResourceGroupName -ServerName $ServerName -Name $AgentName
         }
 
-        Write-PSFMessage -Level Verbose -Message ('Environment check for agent ''{0}'': server {1}, database {2}, agent {3}.' -f
+        Write-PSFMessage -Level Output -Message ('Environment check for agent ''{0}'': server {1}, database {2}, agent {3}.' -f
             $AgentName,
             $(if ($null -ne $server) { 'present' } else { 'absent' }),
             $(if ($null -ne $database) { 'present' } else { 'absent' }),